+++ package.json
@@ -34,6 +34,7 @@
     "@eva-design/eva": "^2.0.0",
     "@eva-design/material": "^2.0.0",
     "@expo/webpack-config": "^0.12.58",
+    "@ptomasroos/react-native-multi-slider": "^2.2.2",
     "@react-native-async-storage/async-storage": "^1.13.4",
     "@react-native-community/masked-view": "^0.1.10",
     "@react-native-picker/picker": "^1.16.3",
