+++ src/myapp/components/products/product-list.tsx
@@ -1,6 +1,6 @@
 import {List} from '@ui-kitten/components';
 import React from 'react';
-import {Image, Platform, StyleSheet, View} from 'react-native';
+import {ScrollView, StyleSheet, View} from 'react-native';
 // Global Styles.
 import globalColors from '../../styles/colors';
 // Hooks.
@@ -10,29 +10,13 @@ import CustomSpinner from '../custom-spinner';
 import ProductCard from './product-card';
 // Types.
 import {ProductListProps} from '../../types/components/products';
-import DefaultText from '../texts/default-text';
-import TitleHeader from '../texts/title-header';
-import globalVars from '../../styles/vars';
+import ProductListEmpty from './product-list-empty';
 
 const ProductList = (props: ProductListProps): React.ReactElement => {
   const {data, isLoading} = useProductsList(
     props.categoryId ?? '',
+    props.name ?? '',
     props.brandId ?? '',
-  );
-
-  const emptyListComponent = () => (
-    <View style={styles.emptyContainer}>
-      <Image
-        style={styles.emptyImage}
-        source={require('./assets/no-results.png')}
-      />
-      <TitleHeader style={styles.emptyTitle}>
-        No se encontraron resultados
-      </TitleHeader>
-      <DefaultText style={styles.emptySubtitle}>
-        Intenta cambiar las opciones de filtros para obtener mejores resultados.
-      </DefaultText>
-    </View>
   );
 
   return isLoading ? (
@@ -42,13 +26,13 @@ const ProductList = (props: ProductListProps): React.ReactElement => {
   ) : (
     <List
       data={data ? data.data : []}
-      ListEmptyComponent={emptyListComponent}
+      ListEmptyComponent={<ProductListEmpty />}
       scrollEnabled={data?.data?.length ? true : false}
       renderItem={({item}) => (
         <ProductCard
           brand={item.brand.name}
-          name={item.name}
           cover_image={item.cover_image}
+          name={item.name}
         />
       )}
       style={styles.container}
@@ -64,22 +48,6 @@ const styles = StyleSheet.create({
   container: {
     backgroundColor: globalColors.backgroundDefault,
     marginTop: 16,
-  },
-  emptyContainer: {
-    padding: globalVars.outsidePadding,
-  },
-  emptyImage: {
-    width: '100%',
-    maxHeight: 250,
-    resizeMode: 'contain',
-  },
-  emptyTitle: {
-    textAlign: 'center',
-  },
-  emptySubtitle: {
-    textAlign: 'center',
-    fontFamily: globalVars.fontBold,
-    fontWeight: Platform.OS === 'ios' ? 'bold' : 'normal',
   },
 });
