+++ src/myapp/hooks/products/useProductsList.ts
@@ -1,10 +1,10 @@
 import {useQuery} from 'react-query';
 import api from '../../services/app-services';
 
-const useProductsList = (categoryId, brandId) =>
-  useQuery(['products', categoryId, brandId], () =>
+const useProductsList = (categoryId = '', name = '', brandId = '') =>
+  useQuery(['products', categoryId, name, brandId], () =>
     api.get(
-      `api/v1/products/?brand=${brandId}&category=${categoryId}&is_active=true`,
+      `api/v1/products/?brand=${brandId}&category=${categoryId}&is_active=true&name=${name}`,
       true,
     ),
   );
