--- src/myapp/layouts/cart/product-list/extra/icons.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { ImageStyle } from 'react-native';
-import { Icon, IconElement } from '@ui-kitten/components';
-
-export const CartIcon = (style: ImageStyle): IconElement => (
-  <Icon {...style} name='shopping-cart'/>
-);
