+++ src/myapp/layouts/products/index.tsx
@@ -1,7 +1,135 @@
-import React, {useLayoutEffect, useEffect, useState} from 'react';
-// My components
-import DefaultText from '../../components/texts/default-text';
+import React, {useState} from 'react';
+import {StyleSheet, TouchableOpacity, View} from 'react-native';
+import {List} from '@ui-kitten/components';
+// Hooks.
+import useGetCategories from '../../hooks/categories/useGetCategories';
+// My Components.
+import AnchorText from '../../components/texts/anchor-text';
+import CustomSpinner from '../../components/custom-spinner';
+import DefaultLayout from '../../components/layouts/default-layout';
+import ProductList from '../../components/products/product-list';
+import SearchInput from '../../components/inputs/search-input';
+import TitleHeader from '../../components/texts/title-header';
+// Global Styles.
+import globalColors from '../../styles/colors';
+import globalVars from '../../styles/vars';
 
-export default (): React.ReactElement => {
-  return <DefaultText children="Hola Mundo" />;
+export default ({navigation, route}): React.ReactElement => {
+  const {data, isLoading} = useGetCategories();
+  const [category, setCategory] = useState<string>('');
+  const [name, setName] = useState<string>('');
+
+  const togglecategory = (id) => {
+    if (category === id) setCategory('');
+    else setCategory(id);
+  };
+
+  return isLoading ? (
+    <CustomSpinner />
+  ) : (
+    <DefaultLayout style={styles.container}>
+      <View style={styles.headerContainer}>
+        <TitleHeader>Productos para tus mascotas</TitleHeader>
+        <View style={styles.filterSection}>
+          <SearchInput onFilter={setName} style={styles.searchInput} />
+          <AnchorText
+            onPress={() =>
+              navigation.navigate('ProductFilter', {
+                category,
+                name,
+              })
+            }
+            style={styles.filterButton}>
+            Filtrar
+          </AnchorText>
+        </View>
+      </View>
+      <View>
+        <List
+          data={data ? data.data : []}
+          horizontal={true}
+          renderItem={({item, index}) => (
+            <TouchableOpacity
+              activeOpacity={0.8}
+              onPress={() => togglecategory(item.id)}
+              style={[
+                styles.filterOption,
+                category === item.id && styles.filterOptionEnabled,
+                index === 0 && styles.filterOptionLeftSpacing,
+                index === data?.data.length - 1 &&
+                  styles.filterOptionRightSpacing,
+              ]}>
+              <TitleHeader
+                style={[
+                  styles.filterOptionText,
+                  category === item.id && styles.filterOptionTextEnabled,
+                ]}>
+                {item.name}
+              </TitleHeader>
+            </TouchableOpacity>
+          )}
+          showsHorizontalScrollIndicator={false}
+          style={styles.filterOptionsContainer}
+        />
+      </View>
+      <View style={styles.resultSection}>
+        <ProductList
+          categoryId={category}
+          name={name}
+          brandId={route.params.brand}
+        />
+      </View>
+    </DefaultLayout>
+  );
 };
+
+const styles = StyleSheet.create({
+  container: {
+    paddingHorizontal: 0,
+  },
+  headerContainer: {
+    paddingHorizontal: globalVars.outsidePadding,
+  },
+  filterSection: {
+    flexDirection: 'row',
+    marginTop: 18,
+  },
+  searchInput: {
+    flexGrow: 1,
+  },
+  filterButton: {
+    paddingLeft: 18,
+    paddingVertical: 18,
+  },
+  filterOptionsContainer: {
+    marginTop: 16,
+    backgroundColor: globalColors.backgroundDefault,
+  },
+  filterOption: {
+    paddingVertical: 6,
+    paddingBottom: 2,
+    paddingHorizontal: 16,
+    backgroundColor: globalColors.backgroundDefault,
+    borderRadius: 10,
+  },
+  filterOptionEnabled: {
+    backgroundColor: globalColors.greenSecondary,
+  },
+  filterOptionText: {
+    color: globalColors.lightGray,
+    fontSize: 16,
+  },
+  filterOptionTextEnabled: {
+    color: globalColors.white,
+  },
+  filterOptionLeftSpacing: {
+    marginLeft: globalVars.outsidePadding,
+  },
+  filterOptionRightSpacing: {
+    marginRight: globalVars.outsidePadding,
+  },
+  resultSection: {
+    flexGrow: 1,
+    flexBasis: 300,
+  },
+});
