--- src/myapp/scenes/cart/product-list.component.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import ContentView from '../../layouts/cart/product-list/index';
-
-export const ProductListScreen = ({navigation}): React.ReactElement => {
-  return <ContentView navigation={navigation} />;
-};
