+++ src/myapp/scenes/products/product-filter.component.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import ContentView from '../../layouts/products/filter/index';
+
+export const ProductFilterScreen = ({
+  navigation,
+  route,
+}): React.ReactElement => <ContentView navigation={navigation} route={route} />;
