+++ src/myapp/scenes/products/product-list.component.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ContentView from '../../layouts/products/index';
+
+export const ProductListScreen = ({navigation, route}): React.ReactElement => (
+  <ContentView navigation={navigation} route={route} />
+);
