+++ src/myapp/types/components/products.ts
@@ -1,6 +1,7 @@
 export interface ProductListProps {
   categoryId?: string;
   brandId?: string;
+  name?: string;
 }
 
 export interface ProductCardProps {
